refactoring: msw 폴더 이동
