+++ .prettierrc
@@ -1,7 +1,11 @@
 {
-  "semi": true,
-  "singleQuote": true,
-  "trailingComma": "es5",
-  "tabWidth": 2,
-  "printWidth": 80
+  "printWidth": 80,        
+  "tabWidth": 2,            
+  "useTabs": false,         
+  "semi": true,             
+  "singleQuote": true,      
+  "trailingComma": "none",   
+  "bracketSpacing": true,   
+  "arrowParens": "always",  
+  "endOfLine": "lf"         
 }
