+++ src/features/user/test/index.ts
@@ -1,0 +1,1 @@
+export * from './handler';
